backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.3
python-multipart>=0.0.12
pydantic-settings>=2.6.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
